Update main.sh
Added validation for the download of the elementary OS system files. If already downloaded, skip the download process

+++ main.sh
@@ -168,8 +168,12 @@ fi
 log_msg "INFO" "Importing device $device_model manifest..."
 . $device_manifest
 
-log_msg "INFO" "Downloading elementary OS system files..."
-run_command "curl -o '$eos_sys_archive' -L -O $eos_sys_archive_url"
+if [ ! -e "$eos_sys_archive" ];then
+      log_msg "INFO" "Downloading elementary OS system files..."
+      run_command "curl -o '$eos_sys_archive' -L -O $eos_sys_archive_url"
+else
+      log_msg "INFO" "elementary OS system files are already downloaded...skipping"
+fi
 
 log_msg "INFO" "Validation elementary OS system files archive md5sum..."
 eos_sys_archive_dl_md5=$(md5sum $eos_sys_archive | awk '{print $1}')
